requirements.txt
Pillow>=10.0.0
colorama>=0.4.6